Implemented random spawning and nearby item menu
Item spawning now picks randomly from an alterable, predefined list of possible items.
Nearby items are now displayed on the bottom right. Scrolling up or down changes the selection and pressing F moves the selection to the cart and updates isInPlayerCart.
I would like to have some sort of indicator on where selection is on the list like an outline, but I am unsure on how to do that.